fix: refactor to python 3.12 and make it work again
