+++ mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
-python_version = 3.9
+python_version = 3.12
 plugins = pydantic.mypy
 exclude = (?x)(
         alembic
