+++ src/app/README
@@ -1,0 +1,2 @@
+Здесь должно лежать само приложение, которое использует пакеты pyhass_mqtt и zont
+
