--- src/devices/__init__.py
@@ -1,6 +1,0 @@
-from .device import Device
-
-
-__all__ = [
-    'Device',
-]
