--- src/entities/__init__.py
@@ -1,6 +1,0 @@
-from .entity import Entity
-
-
-__all__ = [
-    'Entity',
-]
