+++ src/pyhass_mqtt/__init__.py
@@ -1,0 +1,2 @@
+from . import entities
+from . import objects
