+++ src/pyhass_mqtt/entities/__init__.py
@@ -1,3 +1,2 @@
 from . import enums
 from . import models
-from .main import *
